Week7_AnatoBot_FinalProject/anato_backend/requirements_backend.txt
fastapi
uvicorn[standard]
python-dotenv
groq
pydantic

